pyproject.toml
[project]
name = "src"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.1",
    "flask-cors>=6.0.1",
    "google-cloud-aiplatform>=1.103.0",
    "google-genai>=1.25.0",
    "requests>=2.32.4",
]
